Implementation of the message protocol between iframes. We add fixtures to Category and Game model to test the play of a game with iframe.
TODO: Vaidate that width and height are present.
TODO: Load game state from database
TODO: Move from the main app view the 'playgame' method to another app that deals the display of the game. Change the urls.py of the gameshop accordingly.
TODO: Remove the registering in admin.py of Games and Categories respectively.